Core/Misc: Merge TC and fixes typo
Core/Units: Fixed setting power fields from regeneration
Core/Misc: Fix some warnings and logic mistakes found by static code analysis
and more

+++ src/server/game/Spells/SpellInfo.cpp
@@ -385,7 +385,7 @@ bool SpellEffectInfo::IsAura() const
 
 bool SpellEffectInfo::IsAura(AuraType aura) const
 {
-   return IsAura() && AuraType(ApplyAuraName) == aura;
+    return IsAura() && AuraType(ApplyAuraName) == uint32(aura);
 }
 
 bool SpellEffectInfo::IsTargetingArea() const
